Add files via upload
EDA, modeling,and evaluation project applying classifiers or regressors on a chosen dataset. Explore, predict, and analyze with accuracy.

For this lecture we will be working with the Titanic Data Set from Kaggle or github. This is a very famous data set and very often is a student's first step in machine learning!

We'll be trying to predict a classification- survival or deceased. Let's begin our understanding of implementing Logistic Regression in Python for classification.

We'll use a "semi-cleaned" version of the titanic data set, if you use the data set hosted directly on Kaggle or github, you may need to do some additional cleaning not shown in this lecture notebook